+++ src/New/Type/ShieldTypeClass.cpp
@@ -71,6 +71,7 @@ void ShieldTypeClass::LoadFromINI(CCINIClass* pINI)
 	this->BreakAnim.Read(exINI, pSection, "BreakAnim");
 	this->HitAnim.Read(exINI, pSection, "HitAnim");
 	this->BreakWeapon.Read<true>(exINI, pSection, "BreakWeapon");
+	this->HitBright.Read(exINI, pSection, "HitFlash");
 
 	this->AbsorbPercent.Read(exINI, pSection, "AbsorbPercent");
 	this->PassPercent.Read(exINI, pSection, "PassPercent");
@@ -113,6 +114,7 @@ void ShieldTypeClass::Serialize(T& Stm)
 		.Process(this->BreakAnim)
 		.Process(this->HitAnim)
 		.Process(this->BreakWeapon)
+		.Process(this->HitBright)
 		.Process(this->AbsorbPercent)
 		.Process(this->PassPercent)
 		.Process(this->AllowTransfer)